Cleaned up and re-organized template customization
