+++ emhub/__init__.py
@@ -140,9 +140,6 @@ def create_app(test_config=None):
                 'name',             # Simple name
                 ]
 
-    if not "TEMPLATE_MAIN" in app.config:
-        app.config["TEMPLATE_MAIN"] = 'main.html'
-
     portalAPI = app.config.get('SLL_PORTAL_API', None)
     if portalAPI is not None:
         from .data.imports.scilifelab import PortalManager
@@ -184,12 +181,6 @@ def create_app(test_config=None):
         kwargs['possible_booking_owners'] = app.dc.get_pi_labs()
         kwargs['possible_operators'] = app.dc.get_possible_operators()
         kwargs['booking_types'] = app.dm.Booking.TYPES
-        kwargs['session_content'] = app.config.get('TEMPLATE_SESSION_CONTENT',
-                                                   'session_content.html')
-        kwargs['session_body'] = app.config.get('TEMPLATE_SESSION_BODY',
-                                                'create_session_form_body.html')
-        kwargs['dashboard_right'] = app.config.get('TEMPLATE_DASHBOARD_RIGHT',
-                                                   'dashboard_right.html')
         kwargs['currency'] = app.dm.get_config('resources').get('currency', '$')
         try:
             display = app.dm.get_config('bookings')['display']
@@ -235,7 +226,7 @@ def create_app(test_config=None):
 
         register_basic_params(kwargs)
 
-        return flask.render_template(app.config['TEMPLATE_MAIN'], **kwargs)
+        return flask.render_template("main.html", **kwargs)
 
     def _redirect(endpoint, **kwargs):
         return flask.redirect(flask.url_for(endpoint, _external=True, **kwargs))
@@ -402,6 +393,10 @@ def create_app(test_config=None):
         if content_template in templates:
             # Render from templates already in EMhub
             register_basic_params(kwargs)
+
+            if app.is_devel:
+                app.logger.debug(f"template: {content_template}, kwargs: {content_kwargs}")
+
             return flask.render_template(content_template, **kwargs)
 
         error = {
