+++ emhub/templates/create_session_form.html
@@ -13,7 +13,7 @@
         </div>
         <!-- End of Modal header -->
 
-        {% include session_body %}
+        {% include "create_session_form_body.html" %}
 
     <!-- Modal footer -->
     <div class="modal-footer">
