+++ emhub/templates/session_details.html
@@ -6,7 +6,7 @@
     {% include 'include_header.html' %}
 
     <div class="col-xl-12 col-lg-12 col-md-12 col-sm-12 col-12">
-        {% include session_content %}
+        {% include "session_content.html" %}
     </div>
     <div class="row col-6">
         <div class="row col-12"><h5 id="reload_label">Reloading in seconds</h5></div>
